Feat: 중고거래 인기 게시글 TOP3 응답용 DTO 구현
